third_party/blink/web_tests/fullscreen/full-screen-iframe-without-allow-attribute-allowed-from-parent-expected.txt

Test entering full screen security restrictions. An iframe without an allow attribute is still permitted to fullscreen if the request comes from the containing document.

To test manually, press any key - the page should enter full screen mode.

TEST(document.webkitFullscreenEnabled) OK
iframe's webkitFullscreenEnabled should be false:
TEST(iframeMessage == 'document.webkitFullscreenEnabled == false') OK
EVENT(webkitfullscreenchange)
SUCCEED - entered full screen!
END OF TEST
